refactor(auth): adjust sign in and sign out
- use kebab-case for files
- wrap signout logic into independet hooks
