+++ src/app/auth/__hooks/index.ts
@@ -1,1 +1,2 @@
-export * from './useSignIn'
+export * from './use-sign-in'
+export * from './use-sign-out'
